+++ persistence/nosql/async/api/src/main/java/org/apache/polaris/nosql/async/Cancelable.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.async;
+package org.apache.polaris.nosql.async;
 
 import java.util.concurrent.CompletionStage;
