+++ persistence/nosql/async/api/src/testFixtures/java/org/apache/polaris/nosql/async/MockAsyncExec.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.async;
+package org.apache.polaris.nosql.async;
 
 import static com.google.common.base.Preconditions.checkState;
